requirements.txt
alembic==1.6.5
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==2.0.1
Flask-JWT-Extended==4.2.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
lockfile==0.12.2
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
PyJWT==2.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.16.0
SQLAlchemy==1.4.18
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==2.0.1
